+++ scripts/fairness/wage_distribution.jl
@@ -7,11 +7,10 @@ let
         @where startswith(i.quantile,"D")
         @where i.currency == "EUR"
         @where i.indic_il == "SHARE"
-        @select {i.quantile, OBS_VALUE = getfield(i,Symbol(TIME_PERIOD))}
+        @select {i.quantile, OBS_VALUE = Float64(getfield(i,Symbol(TIME_PERIOD)))}
         @collect DataFrame
     end
     sp = sortperm([parse(Int,i[2:end]) for i in belgian_distribution.quantile])
-
     bar(belgian_distribution.quantile[sp],belgian_distribution.OBS_VALUE[sp],alpha=0.3, label = "Belgium")
 
 
@@ -22,7 +21,7 @@ let
         @where i.currency == "EUR"
         @where i.indic_il == "SHARE"
         @group i by i.quantile into g
-        @select {quantile = key(g), OBS_VALUE = mean(getproperty(g,Symbol(TIME_PERIOD)))}
+        @select {quantile = key(g), OBS_VALUE = Float64(mean(getproperty(g,Symbol(TIME_PERIOD))))}
         @collect DataFrame
     end
     sp = sortperm([parse(Int,i[2:end]) for i in belgian_distribution.quantile])
